+++ src/mod/am/coh_deam.scala
@@ -13,11 +13,11 @@ class AnalogSyncDetectorIO[T <: Data](gen: T) extends Bundle {
 }
 
 class SDDeAM[T <: Data](gen: T)(
-  carrierFreq: Long,
+  carrierFreq: Int,
   baseFreqLimit: Double,
   phaseWidth: Int,
   lutWidth: Int,
-  clkFreq: Long,
+  clkFreq: Int,
   filterOrder: Int = 64
 )(implicit analog: Analog[T])
     extends Module {
@@ -30,7 +30,7 @@ class SDDeAM[T <: Data](gen: T)(
   io.out    := fir.io.out
 }
 
-class ASDDeAM[T <: Data](gen: T)(baseFreqLimit: Double, clkFreq: Long, filterOrder: Int = 64)(
+class ASDDeAM[T <: Data](gen: T)(baseFreqLimit: Double, clkFreq: Int, filterOrder: Int = 64)(
   implicit analog: Analog[T]
 ) extends Module {
   override def desiredName = s"deam_asd_fl${baseFreqLimit}_o$filterOrder"
